Não permitir remover adms Padrões

+++ routes/api/user.js
@@ -387,9 +387,13 @@ router.post('/forgot-password', async (req, res) => {
 })
 
 router.delete('/:user_id', async (req, res, next) => {
-    await User.findByIdAndDelete(req.params.user_id)
-
-    return res.status(200).json({ message: 'Excluído' })
+    let user =  await User.findOne({_id: req.params.user_id});
+    if((user.email == "admin@admin.com") || (user.email == "admin@planetsweb.com.br")){
+        return res.status(400).json({ message: 'Não é possivel remover um administrador de nível alto criado pelo sistema.' })
+    }else{
+        await User.findByIdAndDelete(req.params.user_id)
+        return res.status(200).json({ message: 'Excluído' })
+    }
 });
 
 router.post('/reset/:token', async (req, res, next) => {
